moved report data specific constants into their own files
