+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/reports/Constants.java
@@ -7,16 +7,7 @@ public final class Constants {
 
     public static final String REPORT_OFFICE = "\"Report_Office\":\"";
     public static final String TOTAL_CASES = "\"Total_Cases\":\"";
-    public static final String REPORT_DATE = "\"Report_Date\":\"";
-    public static final String TOTAL_SINGLE = "\"Total_Single\":\"";
-    public static final String TOTAL_MULTIPLE = "\"Total_Multiple\":\"";
-    public static final String TOTAL_WITHIN_4WEEKS = "\"Total_Within_4Weeks\":\"";
-    public static final String TOTAL_PERCENT_WITHIN_4WEEKS = "\"Total_Percent_Within_4Weeks\":\"";
-    public static final String TOTAL_NOT_WITHIN_4WEEKS = "\"Total_Not_Within_4Weeks\":\"";
-    public static final String TOTAL_PERCENT_NOT_WITHIN_4WEEKS = "\"Total_Percent_Not_Within_4Weeks\":\"";
     public static final String REPORT_DETAILS = "reportDetails";
-    public static final String REPORT_DETAILS_SINGLE = "reportDetailsSingle";
-    public static final String REPORT_DETAILS_MULTIPLE = "reportDetailsMultiple";
 
     public static final String ELASTICSEARCH_FIELD_MANAGING_OFFICE_KEYWORD = "data.managingOffice.keyword";
     public static final String ELASTICSEARCH_FIELD_STATE_KEYWORD = "state.keyword";
