+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/reports/nochangeincurrentposition/NoPositionChangeCcdDataSource.java
@@ -17,7 +17,6 @@ import static uk.gov.hmcts.ecm.common.model.helper.Constants.OLD_DATE_TIME_PATTE
 @RequiredArgsConstructor
 @Slf4j
 public class NoPositionChangeCcdDataSource implements NoPositionChangeDataSource {
-
     private final String authToken;
     private final CcdClient ccdClient;
 
@@ -27,15 +26,14 @@ public class NoPositionChangeCcdDataSource implements NoPositionChangeDataSource
             var reportDate3MonthsAgo = LocalDate.parse(reportDate, OLD_DATE_TIME_PATTERN2).minusMonths(3)
                     .format(OLD_DATE_TIME_PATTERN2);
             var query = NoPositionChangeElasticSearchQuery.create(reportDate3MonthsAgo);
-            var submitEvents = new ArrayList<NoPositionChangeSubmitEvent>();
             var searchResult = ccdClient.runElasticSearch(authToken, caseTypeId, query,
                     NoPositionChangeSearchResult.class);
 
             if (searchResult != null && CollectionUtils.isNotEmpty(searchResult.getCases())) {
-                submitEvents.addAll(searchResult.getCases());
+                return searchResult.getCases();
             }
 
-            return submitEvents;
+            return new ArrayList<>();
         } catch (Exception e) {
             throw new ReportException(String.format(
                     "Failed to get No Change In Current Position search results for case type id %s", caseTypeId), e);
@@ -46,14 +44,13 @@ public class NoPositionChangeCcdDataSource implements NoPositionChangeDataSource
     public List<SubmitMultipleEvent> getMultiplesData(String caseTypeId, List<String> multipleRefsList) {
         try {
             var query = NoPositionChangeMultiplesElasticSearchQuery.create(multipleRefsList);
-            var submitEvents = new ArrayList<SubmitMultipleEvent>();
             var searchResult = ccdClient.runElasticSearch(authToken, caseTypeId, query, MultipleCaseSearchResult.class);
 
             if (searchResult != null && CollectionUtils.isNotEmpty(searchResult.getCases())) {
-                submitEvents.addAll(searchResult.getCases());
+                return searchResult.getCases();
             }
 
-            return submitEvents;
+            return new ArrayList<>();
         } catch (Exception e) {
             throw new ReportException(String.format(
                     "Failed to get No Change In Current Position multiples search results for case type id %s",
