+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/ReportDataService.java
@@ -106,7 +106,7 @@ public class ReportDataService {
         }
     }
 
-    private void setSharedReportDocumentFields(ListingData reportData, ListingDetails listingDetails, 
+    private void setSharedReportDocumentFields(ListingData reportData, ListingDetails listingDetails,
                                                boolean isDateRangeReport) {
         reportData.setDocumentName(listingDetails.getCaseData().getDocumentName());
         reportData.setReportType(listingDetails.getCaseData().getReportType());
